Description
Part 1 –Create a simple stack class based on the provided stack interface. Implement it using a list of Nodes

Part 2 – Create a program that checks uses a stack to solve the stock span problem.

The stock span problem is a financial problem where we have a series of n daily price quotes for a stock and we need to calculate span of stock’s price for all n days. 

The span 𝑆_𝑖 of the stock’s price on a given day i is defined as the maximum number of consecutive days just before the given day, for which the price of the stock on the current day is less than or equal to its price on the given day.

For example, if an array of 7 days prices is given as:
 {100, 80, 60, 70, 60, 75, 85}, 
then the span values for corresponding 7 days are:
 {1, 1, 1, 2, 1, 4, 6}